+++ docs/changelog.rst
@@ -10,10 +10,13 @@ Unreleased
 ==========
 New
 ---
+- Before writing the forcing to disk, throw an error for any missing values (#748)
 
 Fixed
 -----
 - Fixed `setup_reservoirs_no_control()` to be able to handle both strings and path objects. (#770)
+- Fixed outdated imports in ``hydromt_wflow.components.staticmaps``. (#782)
+- Fixed `hydrography` workflow to correctly remove no data values outside of the model region. (#785)
 
 Removed
 -------